requirements.txt
flask
openai
langchain
tiktoken
requests
python-dotenv